packages/suds-ui/pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "suds-ui"
version = "0.1.0"
description = "Streamlit UI for interacting with SUDS FastAPI."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "SUDS Team" }]

dependencies = [
  "streamlit>=1.30",
  "requests>=2.31",
  "pydeck>=0.8",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/suds_ui"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]